fixed current weather box and fixed pressure
git-svn-id: https://vcs.maemo.org/svn/omweather/trunk@4563 5aa3d57c-301e-0410-a7f2-8fec96c11689

+++ qtomweather/core/data.cpp
@@ -243,8 +243,8 @@ namespace Core {
     }
 ////////////////////////////////////////////////////////////////////////////////
     void
-    Data::Pressure(int pressure){
-        _pressure = pressure;
+    Data::Pressure(float pressure){
+        _pressure =(int)pressure;
     }
 
     int
